requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-admin==2.0.1
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
